bootstrap: Allow executing linux code standalone
